requirements.txt
pillow==10.3.0
PyYAML==6.0.1
